- added the restriction (task management by user) - added exception for AccessDenied - added securityUtil for getting current user - use jpa persistent context (entity tracking) - improved security - improved exception handler - improved mapper classes for injection

+++ src/main/java/com/xeon/todolist/service/TaskService.java
@@ -4,12 +4,15 @@ import com.xeon.todolist.dto.CreateTaskRequest;
 import com.xeon.todolist.dto.TaskResponse;
 import com.xeon.todolist.dto.UpdateTaskRequest;
 import com.xeon.todolist.entity.Users;
+import com.xeon.todolist.exception.TaskAccessDeniedException;
 import com.xeon.todolist.exception.TaskNotFoundException;
 import com.xeon.todolist.exception.UserNotFoundException;
 import com.xeon.todolist.mapper.TaskMapper;
 import com.xeon.todolist.entity.Tasks;
 import com.xeon.todolist.repository.TaskRepository;
 import com.xeon.todolist.repository.UserRepository;
+import com.xeon.todolist.security.UserPrincipal;
+import com.xeon.todolist.util.SecurityUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.data.autoconfigure.web.DataWebProperties;
 import org.springframework.data.domain.Page;
@@ -18,69 +21,88 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
 @Service
+@Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class TaskService {
 
     private final TaskRepository taskRepository;
 
-    private final UserRepository userRepository;
+    private final TaskMapper taskMapper;
 
+    private final SecurityUtil securityUtil;
+
+    @Transactional
     public TaskResponse createTask(CreateTaskRequest createTaskRequest) {
-        String username = Objects.requireNonNull(SecurityContextHolder.getContext().getAuthentication()).getName();
-        Users user = userRepository.findByUsername(username).orElseThrow(() -> new UserNotFoundException("User not found"));
-
-        Tasks task = new Tasks();
+        Users user = securityUtil.getCurrentUser();
+        Tasks task = new Tasks(); // need save() since it creates a new entity
         task.setTitle(createTaskRequest.getTitle());
         task.setPriority(createTaskRequest.getPriority());
         task.setUser(user);
-        return TaskMapper.toResponse(taskRepository.save(task));
+        return taskMapper.toResponse(taskRepository.save(task));
     }
 
-//    public List<TaskResponse> getAllTasks() {
-//        return taskRepository.findAll().stream().map(TaskMapper::toResponse).toList();
-//    }
-
+    @Transactional
     public void deleteTask(long taskId) {
+        Users user = securityUtil.getCurrentUser();
         Tasks task = taskRepository.findById(taskId).orElseThrow(() -> new TaskNotFoundException("Task not found"));
+        if (!task.getUser().getId().equals(user.getId())){
+            throw new TaskAccessDeniedException("Task Access Denied");
+        }
         taskRepository.delete(task);
     }
 
+    @Transactional
     public TaskResponse updateTask(long taskId, UpdateTaskRequest updateTaskRequest) {
         Tasks task = taskRepository.findById(taskId).orElseThrow(() -> new TaskNotFoundException("Task not found"));
+        //check user first
+        Users user = securityUtil.getCurrentUser();
+        if (!task.getUser().getId().equals(user.getId())) {
+            throw new TaskAccessDeniedException("Task Access Denied");
+        }
         task.setTitle(updateTaskRequest.getTitle());
-        task.setId(taskId);
         task.setCompleted(updateTaskRequest.isCompleted());
         task.setPriority(updateTaskRequest.getPriority());
-        return TaskMapper.toResponse(task);
+        return taskMapper.toResponse(task);
     }
 
+    @Transactional
     public TaskResponse toggleTaskStatus(long taskId) {
         Tasks task = taskRepository.findById(taskId).orElseThrow(() -> new TaskNotFoundException("Task not found"));
+        Users user = securityUtil.getCurrentUser();
+        if (!task.getUser().getId().equals(user.getId())) {
+            throw new TaskAccessDeniedException("Task Access Denied");
+        }
         task.setCompleted(!task.isCompleted());
-        return TaskMapper.toResponse(task);
+        return taskMapper.toResponse(task);
     }
 
     public TaskResponse getTask(long taskId) {
         Tasks task = taskRepository.findById(taskId).orElseThrow(() -> new TaskNotFoundException("Task not found"));
-        return TaskMapper.toResponse(task);
+        Users user = securityUtil.getCurrentUser();
+        if (!task.getUser().getId().equals(user.getId())) {
+            throw new TaskAccessDeniedException("Task Access Denied");
+        }
+        return taskMapper.toResponse(task);
     }
 
     public Page<TaskResponse> getAllTasksByPage(int page, int size, List<String> sort) {
 
+        Users user = securityUtil.getCurrentUser();
         Pageable pageable = PageRequest.of(page, size, parseSort(sort));
 
-        return taskRepository.findAll(pageable).map(TaskMapper::toResponse);
+        return taskRepository.findTasksByUsername(user.getUsername(), pageable).map(taskMapper::toResponse);
     }
 
 
     public Sort parseSort(List<String> sortParam) {
-        if (sortParam.isEmpty()) {
+        if (sortParam == null || sortParam.isEmpty()) {
             return Sort.by(Sort.Direction.DESC, "priorityWeight");
         }
